refactor(fab): restructure and update typing for Fab component
Relocated Fab component files to new directories for better organization and maintainability. Updated typings to use the Component utility for improved type safety and consistency across the codebase. Adjusted imports, styles, and interfaces to align with the new structure and typing changes.
