+++ src/components/action/Fab.tsx
@@ -1,11 +1,12 @@
 import React, { useRef } from 'react';
-import { Icon } from '../../../icon';
+import { Icon } from '../../icon';
 
-import { RippleEffect } from '../../../effects/ripple';
+import { RippleEffect } from '../../effects/ripple';
 import { AnimatePresence, motion } from 'framer-motion';
-import { FabProps } from './fab.interface';
-import { fabStyle } from './fab.style';
-import { classNames } from '../../../utils';
+import { FabInterface } from '../../interfaces/fab.interface';
+import { fabStyle } from '../../styles/fab.style';
+import { classNames } from '../../utils';
+import { ReactProps } from '../../utils/component';
 
 export const Fab = ({
   className,
@@ -18,7 +19,7 @@ export const Fab = ({
   isExtended = false,
   ref,
   ...restProps
-}: FabProps) => {
+}: ReactProps<FabInterface>) => {
   const ElementType = href ? 'a' : 'button';
 
   const styles = fabStyle({
